Backend további átrendezése és barátok végpont javítása
