+++ backend/api/blocks.js
@@ -1,15 +1,7 @@
 import express from "express";
-import * as service from "../service/blocks.js";
+import * as controller from "../controller/blocks.js";
 import * as validator from "../validator/block.js";
-import {
-  upload,
-  createResponse,
-  authenticate,
-  modifyTargetUser,
-  handleCaughtError,
-  handleSequelizeUniqueConstraintError,
-  isSequelizeUniqueConstraintError,
-} from "../common/common.js";
+import { upload, authenticate, modifyTargetUser } from "../common/common.js";
 
 const router = express.Router();
 
@@ -18,27 +10,7 @@ router.get(
   authenticate(),
   modifyTargetUser(),
   validator.POST,
-  async (request, response) => {
-    if (!request.valid) {
-      return response
-        .status(400)
-        .json(createResponse(false, null, "Invalid user ID"));
-    }
-
-    try {
-      const result = await service.getBlockedUsers({
-        blockerId: request.targetUser.id,
-      });
-
-      response
-        .status(200)
-        .json(
-          createResponse(true, result, "Blocked users retrieved successfully"),
-        );
-    } catch (error) {
-      handleCaughtError(response, error);
-    }
-  },
+  controller.getBlockedUsers,
 );
 
 router.post(
@@ -47,32 +19,7 @@ router.post(
   modifyTargetUser(),
   upload.none(),
   validator.POST,
-  async (request, response) => {
-    if (!request.valid) {
-      return response
-        .status(400)
-        .json(createResponse(false, null, "Invalid user ID"));
-    }
-
-    try {
-      await service.blockUser({
-        blockerId: request.targetUser.id,
-        blockedId: request.body.userId,
-      });
-
-      response
-        .status(201)
-        .json(createResponse(true, null, "User blocked successfully"));
-    } catch (error) {
-      if (isSequelizeUniqueConstraintError(error)) {
-        return handleSequelizeUniqueConstraintError(
-          response,
-          "This user has already been blocked",
-        );
-      }
-      handleCaughtError(response, error);
-    }
-  },
+  controller.blockUser,
 );
 
 router.delete(
@@ -81,26 +28,7 @@ router.delete(
   modifyTargetUser(),
   upload.none(),
   validator.POST,
-  async (request, response) => {
-    if (!request.valid) {
-      return response
-        .status(400)
-        .json(createResponse(false, null, "Invalid user ID"));
-    }
-
-    try {
-      await service.unblockUser({
-        blockerId: request.targetUser.id,
-        blockedId: request.body.userId,
-      });
-
-      response
-        .status(200)
-        .json(createResponse(true, null, "User unblocked successfully"));
-    } catch (error) {
-      handleCaughtError(response, error);
-    }
-  },
+  controller.unblockUser,
 );
 
 export default router;
