+++ backend/api/reactions.js
@@ -1,14 +1,12 @@
 import express from "express";
-import * as service from "../service/reactions.js";
+import * as controller from "../controller/reactions.js";
 import {
   upload,
-  createResponse,
   authenticate,
-  handleExpressValidatorErrors,
-  handleCaughtError,
   modifyTargetUser,
+  handleValidation,
 } from "../common/common.js";
-import { checkSchema, validationResult } from "express-validator";
+import { checkSchema } from "express-validator";
 import * as validator from "../validator/reaction.js";
 
 const router = express.Router();
@@ -18,25 +16,8 @@ router.get(
   authenticate(),
   modifyTargetUser(),
   checkSchema(validator.GET),
-  async (request, response) => {
-    const errors = validationResult(request);
-
-    if (!errors.isEmpty())
-      return handleExpressValidatorErrors(response, errors);
-
-    try {
-      const result = await service.getUserReaction({
-        userId: request.targetUser.id,
-        targetId: request.params.targetId,
-      });
-
-      response
-        .status(200)
-        .json(createResponse(true, result, "Reaction fetched"));
-    } catch (error) {
-      handleCaughtError(response, error);
-    }
-  },
+  handleValidation,
+  controller.get,
 );
 
 // prettier-ignore
@@ -46,26 +27,8 @@ router.post(
   modifyTargetUser(),
   upload.none(),
   checkSchema(validator.POST),
-  async (request, response) => {
-    const errors = validationResult(request);
-
-    if (!errors.isEmpty())
-      return handleExpressValidatorErrors(response, errors);
-
-    try {
-      const result = await service.createUserReaction({
-        userId: request.targetUser.id,
-        targetId: request.body.targetId,
-        reactionType: request.body.type
-      });
-
-      response
-        .status(200)
-        .json(createResponse(true, result, "Reaction added or updated"));
-    } catch (error) {
-      handleCaughtError(response, error);
-    }
-  },
+  handleValidation,
+  controller.create,
 );
 
 export default router;
